Remove estimate gas module
This is no longer needed now that web3's CallRequest allows setting `to`
to None.
